add negative test for min gas price
