+++ fuel-tests/tests/tx.rs
@@ -1,5 +1,4 @@
 use chrono::Utc;
-use fuel_core::chain_config::{ChainConfig, CoinConfig, ContractConfig, StateConfig};
 use fuel_core::executor::ExecutionMode;
 use fuel_core::model::{FuelBlock, FuelBlockHeader};
 use fuel_core::{
@@ -7,14 +6,14 @@ use fuel_core::{
     executor::Executor,
     service::{Config, FuelService},
 };
-use fuel_crypto::SecretKey;
 use fuel_gql_client::client::types::TransactionStatus;
 use fuel_gql_client::client::{FuelClient, PageDirection, PaginationRequest};
-use fuel_tx::TransactionBuilder;
 use fuel_vm::{consts::*, prelude::*};
 use itertools::Itertools;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use std::io;
+
+mod utxo_validation;
 
 #[test]
 fn basic_script_snapshot() {
@@ -122,145 +121,6 @@ async fn submit() {
         .unwrap()
         .transaction;
     assert_eq!(tx.id(), ret_tx.id());
-}
-
-#[tokio::test]
-async fn submit_utxo_verified_tx_with_min_gas_price() {
-    let mut rng = StdRng::seed_from_u64(2322);
-
-    let test_contract_code = vec![];
-    let test_contract = Contract::from(test_contract_code.clone());
-    let root = test_contract.root();
-    let test_contract_salt: Salt = rng.gen();
-    let contract_id = test_contract.id(
-        &test_contract_salt.clone(),
-        &root,
-        &Contract::default_state_root(),
-    );
-
-    // initialize transactions
-    let transactions = (1..10 + 1)
-        .into_iter()
-        .map(|i| {
-            let secret = SecretKey::random(&mut rng);
-            TransactionBuilder::script(
-                Opcode::RET(REG_ONE).to_bytes().into_iter().collect(),
-                vec![],
-            )
-            .gas_limit(100)
-            .gas_price(1)
-            .byte_price(1)
-            .add_unsigned_coin_input(
-                rng.gen(),
-                &secret,
-                1000 + i,
-                Default::default(),
-                0,
-                vec![],
-                vec![],
-            )
-            .add_input(Input::Contract {
-                utxo_id: Default::default(),
-                balance_root: Default::default(),
-                state_root: Default::default(),
-                contract_id,
-            })
-            .add_output(Output::Change {
-                amount: 0,
-                asset_id: Default::default(),
-                to: rng.gen(),
-            })
-            .add_output(Output::Contract {
-                input_index: 1,
-                balance_root: Default::default(),
-                state_root: Default::default(),
-            })
-            .finalize()
-        })
-        .collect_vec();
-
-    // setup genesis block with coins that transactions can spend
-    let genesis_coins = transactions
-        .iter()
-        .flat_map(|t| t.inputs())
-        .filter_map(|input| {
-            if let Input::Coin {
-                amount,
-                owner,
-                asset_id,
-                utxo_id,
-                ..
-            } = input
-            {
-                Some(CoinConfig {
-                    tx_id: Some(*utxo_id.tx_id()),
-                    output_index: Some(utxo_id.output_index() as u64),
-                    block_created: None,
-                    maturity: None,
-                    owner: *owner,
-                    amount: *amount,
-                    asset_id: *asset_id,
-                })
-            } else {
-                None
-            }
-        })
-        .collect_vec();
-
-    let config = Config {
-        utxo_validation: true,
-        tx_pool_config: fuel_txpool::Config {
-            min_byte_price: 1,
-            min_gas_price: 1,
-            ..Default::default()
-        },
-        chain_conf: ChainConfig {
-            initial_state: Some(StateConfig {
-                coins: Some(genesis_coins),
-                contracts: Some(vec![ContractConfig {
-                    code: test_contract_code,
-                    salt: test_contract_salt,
-                    state: None,
-                    balances: None,
-                }]),
-                ..StateConfig::default()
-            }),
-            ..ChainConfig::local_testnet()
-        },
-        ..Config::local_node()
-    };
-
-    let srv = FuelService::new_node(config).await.unwrap();
-    let client = FuelClient::from(srv.bound_address);
-
-    for tx in transactions {
-        let id = client.submit(&tx).await.unwrap();
-        // verify that the tx returned from the api matches the submitted tx
-        let ret_tx = client
-            .transaction(&id.0.to_string())
-            .await
-            .unwrap()
-            .unwrap()
-            .transaction;
-
-        let transaction_result = client
-            .transaction_status(&ret_tx.id().to_string())
-            .await
-            .ok()
-            .unwrap();
-
-        if let TransactionStatus::Success { block_id, .. } = transaction_result.clone() {
-            let block_exists = client.block(&block_id).await.unwrap();
-
-            assert!(block_exists.is_some());
-        }
-
-        // Once https://github.com/FuelLabs/fuel-core/issues/50 is resolved this should rely on the Submitted Status rather than Success
-        assert!(matches!(
-            transaction_result,
-            TransactionStatus::Success { .. }
-        ));
-    }
 }
 
 #[ignore]
